[quality] consolidate lint rule implementations
This consolidates the lint rule implementations for clang_format, gofmt,
and protolint, so a common rule implementation and template script is
used for all lint rules, which operate in a similar fashion.

Additionally, this updates the clang_format lint rule to use the
clang_format binary included in the lowRISC RV32 toolchain, for
improved build reproducibility.

Signed-off-by: Tim Trippel <ttrippel@google.com>