[tests] and different operation modes to integration testing script
Previously, `./run_integration_tests.sh` would build a release tarball
of all the provisioning binaries, along with container images containing
each provisioning binary, and then unpack the binaries into the deployment
staging area, even if only the containers were run. This was slow when
debugging because compressing and uncompressing binaries (that are not
even used) is slow.

This adds a `-c` mode to the `./run_integration_tests.sh` to only build
an deploy the container images to save time when debugging.

Additionally, a `-d` (debug) mode is added that prevents the containers
from being torn down on script exit to enable debugging test failures.

Signed-off-by: Tim Trippel <ttrippel@google.com>