[container] build container with SoftHSM2 from Dockerfile
This adds a Dockerfile, build script, and Bazel build rules to build a
container with SoftHSM2 inside (from a Dockerfile) so that it may be used
by the SPM server container.

After being built, the container was then pushed to the public OpenTitan
Provisioning container registry, found here:
us-docker.pkg.dev/opentitan/opentitan-public

This is required fix the integration tests since the previous base
container image used by the SPM server could contain a version of the
PCKS#11 shared library that was not built in the same environment the
SPM server was built it, causing dependency issues.

Signed-off-by: Tim Trippel <ttrippel@google.com>