[pa,auth,loadtest] update PA to accept userID from client
This updates the PA and auth_controller to accept a userID from the
client instead of always defaulting to using the client IP address as
the user ID. This way we can test several sessions eminating from the
same host.

Additionally, this updates the loadtest to send a userID to the PA
upon session initialization and closing.

With this change, we can now execute the load test with several clients
running on the same host machine (with the same IP address).

Signed-off-by: Tim Trippel <ttrippel@google.com>