[test] Configure hstmtool in dev environment.
This change adds the `hsmtool` binary to the `//release` target, and
updates the `config/dev/deploy.sh` script to deploy the binary in the
development environment.

In addition to this, the integration test infrastructure was updated to
initialize a portion of the `sival` sku by invoking `hsmtool` with an
import key script.

Signed-off-by: Miguel Osorio <miguelosorio@google.com>