[ate] Add `EndorseCerts` call.
This change adds the `EndorseCerts` function to the `ate_api.h` library.
The function calls the provisioning appliance backend to endorse
certificates with a given key. The function performs the conversion from
C++ to C data structures.

The `ate_dll` implementation uses C data structures to be able to be
integrated with older compilers with limited C++ support.

Signed-off-by: Miguel Osorio <miguelosorio@google.com>