[pk11] Remove support for HSMType.
The HSM supported attributes and mechanisms should be part of the SKU
configuration. This change removes the `pl11.HSMType` in favor of
applying attributes via `pk11.KeyOptions`, and by defining the key wrap
mechanism via a `skucfg.Config{AttrName}` setting.

In a follow up PR, we are planning to add SKU configurations for the
Luna HSM.

Signed-off-by: Miguel Osorio <miguelosorio@google.com>