[hsm] Add Thales test configuration.
This change adds Thales HSM configuration scripts required to run
integration tests for the `sival` SKU. The set of keys provisioned are
equivalent to the ones used in the SoftHSM configuration.

Thales HSM uses a different set of mechanisms for key wrap/unwrap. Such
mechanisms are selected in the configuration scripts.

Signed-off-by: Miguel Osorio <miguelosorio@google.com>