Create BatchRegisterDevice method and messages
This is in preparation to support BatchRegister in both
ProxyBufferServer and external registries. With this feature we can now register multiple devices with a single call.

I also added an RPC error message to the response field, so we can set
individual failures for each request in a batch request.