[spm] Implement WAS signature verification.
The Wafer Authentication Secret (WAS) is injected during `CP` testing into a
write-only info flash page. Read access is later enabled in the `FT`
stage when the device transitions to `PROD` state.

During personalization, the device uses the WAS to sign the `tbs`
certificates sent to the provisioning appliance in the `EndorseCerts`
RPC. The provisioning appliance verifies the signature over the `tbs` by
recreating the WAS. This operation is used to authorize the certificate
endorsement operation.

Signed-off-by: Miguel Osorio <miguelosorio@google.com>