Change HTTP registry to read headers from file on each request
Since it's possible that the headers are not static (typically
authorization tokens have a lifetime), we need a way of ensuring these
headers can be changed without the need to restart the service or
implementation details.

By reading a header file on every request, we have a way of updating
said headers without requiring to restart the job every time. It does
take a toll on request processing time, but if the file is small (I
wouldn't expect more than 20 headers), it should be fine.