[spm] Use diversifier instead of device_id in EndorseCerts
This change updates the EndorseCerts API to use a diversifying string
instead of a device identifier in the Wafer Authentication Secret (WAS)
key derivation.

This is to align the key derivation function with the token generation
flow used in the ATE DLL library, which constaints all diversification
strings to be less than or equal to 48 bytes, which 0 padding up to the
maximum size.

Signed-off-by: Miguel Osorio <miguelosorio@google.com>