Allow ProxyBuffer syncer to return fatal error on retry failure
This commit allows us to have a maximum numbers of retries per device.
If the device is retried more than said number, it will cause a fatal
error, which is propagated via FatalErrors() channel.

I also fixed a bug in the Stop() logic by creating the channel with a
size 1.