[spm] Refactor SKU management.
Move SKU configuration and SE initialization into a separate module
(skumgr.go). This enables the loading of a SKU into an SE instance
without having to initialize the SPM gRPC service, allowing the SKU and
HSM connection to be used in test and utility programs.

The skumgr connection will be used in a follow up commit to implement a
TBS certificate generation utility which will enable loadtest and
unittests of across all SKU configurations.

Signed-off-by: Miguel Osorio <miguelosorio@google.com>