[spm] Implement TBS generation utility.
The `tbsgen.go` package provides the ability to generate valid
To-Be-Signed certificates for any of the SKU signing keys. This is done
by signing the certificate with the SKU key and then removing it before
returning to the caller. It has to be done this way to be able to
leverage the crypto/x509 certificate generation implementation.

Signed-off-by: Miguel Osorio <miguelosorio@google.com>