[spm] Export ICA SubjectKeyId
This replaces the `GetCaSerialNumbers` for `GetCaSubjecKeys`. This is to
allow Intermediate CA (ICA) certificates to have a serial number set to
a value different than its SubjectKeyId. openssl doesn't set the serial
number to the SubjectKeyId by default.

The ICA SubjectKeyId is used by the DUT to populate the AuthorityKeyID.
Below is a sample SPM log after applying the change in this commit:

```console
2025/06/08 20:43:14 DUT cert AuthorityKeyID: ff12c119c25880fefa2ff765551aef255c0c5622
2025/06/08 20:43:14 DUT cert Issuer: CN=Google Engineering ICA,OU=Engineering,O=Google,ST=CA,C=US
2025/06/08 20:43:14 ICA SubjectKeyID: ff12c119c25880fefa2ff765551aef255c0c5622
2025/06/08 20:43:14 ICA Subject: CN=Google Engineering ICA,OU=Engineering,O=Google,ST=CA,C=US
```

Signed-off-by: Miguel Osorio <miguelosorio@google.com>