[bazel] bump OT repo version
This bumps the lowrisc_opentitan bazel repo version, and subsequently
the OpenTitan provisioning firmware and bitstream dependencies.

Additionally, this adds a CRC computation to the RMA unlock token
message as the latest version of the personalization firmware now
requires the host to send a valid CRC with the RMA unlock token JSON
message.

Signed-off-by: Tim Trippel <ttrippel@google.com>