[spm]: Update EndorseData public key format
The `se_pk11` package was returning a raw ECDSA public key, which caused
parsing failures in the SPM. This change updates the implementation to
return the public key in the standard `SubjectPublicKeyInfo` format
by using `x509.MarshalPKIXPublicKey`.

This fixes ASN.1 parsing errors and ensures interoperability. The
updated key format is now correctly handled when endorsing data and
generating certificates.

This change also updates the EndorseData parameters to use P256-SHA256
instead of P384-SHA384. This is to match the SPM identity key
configuration used in the HSMs.

Signed-off-by: Miguel Osorio <miguelosorio@google.com>