[ate] rename dut_spi_frame_t struct
The size of a SPI console frame differs depending on whether the
direction is TX or RX. The current SPI frame struct is the size of a
DUT TX frame (one the DUT sends to the ATE). In order to add a new SPI
frame type for the RX direction, we must rename the existing SPI frame
struct.

Additionally, this renames `cursor` field to `size` in the newly renamed
`dut_tx_spi_frame_t` struct. The cursor field has always been treated as
the size of the SPI frame, so we rename it as such.

Signed-off-by: Tim Trippel <ttrippel@google.com>